+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Upcoming release
 
+### Features
+
+- Adding `validateSource` to `HtmlValidate` allowing to manually passing a
+  source.
+
 ## 0.16.1 (2018-12-16)
 
 ### Bugfixes